Changed how projects for testing are stored in the test plugin. Added TypeMnemonicTest with test for the project reassignment method in TypeMnemonic.

+++ plugin.tests/src/randoop/plugin/tests/resources/FileResources.java
@@ -10,26 +10,24 @@ import java.io.OutputStream;
 import junit.framework.TestCase;
 
 public class FileResources extends TestCase {
-  public static void copy(File source, File destination) throws IOException {
-    assertTrue(source.exists());
-
+  public static void copy(File source, File destination, boolean noClobber) throws IOException {
     if (source.isFile()) {
-      copyFile(source, destination);
+      copyFile(source, destination, noClobber);
     } else {
-      copyDirectory(source, destination);
+      copyDirectory(source, destination, noClobber);
     }
   }
 
-  private static void copyDirectory(File source, File destination) throws IOException {
+  private static void copyDirectory(File source, File destination, boolean noClobber) throws IOException {
     assertTrue(source.isDirectory());
     assertTrue(!destination.exists() || destination.isDirectory());
     
-    if (destination.exists()) {
+    if (!destination.exists()) {
       destination.mkdirs();
     }
 
     for (File f : source.listFiles()) {
-      copy(f, new File(destination, f.getName()));
+      copy(f, new File(destination, f.getName()), noClobber);
     }
   }
 
@@ -41,16 +39,21 @@ public class FileResources extends TestCase {
    *          file to be copied
    * @param destination
    *          location of new file
+   * @param noClobber 
    * @param monitor
    * @throws IOException
    */
-  private static void copyFile(File source, File destination) throws IOException {
-    assertFalse(destination.exists());
-        
-    // Create the directories and file for destination
-    destination.getParentFile().mkdirs();
-    destination.createNewFile();
-
+  private static void copyFile(File source, File destination, boolean noClobber) throws IOException {
+    if (noClobber && destination.exists()) {
+      return;
+    }
+    
+    if (!destination.exists()) {
+      // Create the directories and file for destination
+      destination.getParentFile().mkdirs();
+      destination.createNewFile();
+    }
+    
     InputStream in = new FileInputStream(source);
     OutputStream out = new FileOutputStream(destination);
